requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioflask==0.4.0
aiohttp==3.7.4.post0; python_version >= '3.6'
asgiref==3.5.2
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cachecontrol==0.12.11; python_version >= '3.6'
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.6.15; python_version >= '3.6'
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
charset-normalizer==2.1.0; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
discord.py==1.7.3; python_full_version >= '3.5.3'
discord==1.7.3
firebase-admin==5.2.0
flask-cors==3.0.10
flask[async]==2.1.2
google-api-core[grpc]==2.8.2; platform_python_implementation != 'PyPy'
google-api-python-client==2.52.0; python_version >= '3.6'
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2; python_version >= '3.6'
google-auth==2.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-core==2.3.1; python_version >= '3.6'
google-cloud-firestore==2.5.3; platform_python_implementation != 'PyPy'
google-cloud-storage==2.4.0; python_version >= '3.7'
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.3.3; python_version >= '3.6'
googleapis-common-protos==1.56.3; python_version >= '3.6'
greenlet==1.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
greenletio==0.9.0; python_version >= '3.6'
grpcio-status==1.47.0
grpcio==1.47.0
gspread==5.4.0
h11==0.13.0; python_version >= '3.6'
httplib2==0.20.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3.5'
importlib-metadata==4.12.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
msgpack==1.0.4
multidict==6.0.2; python_version >= '3.7'
oauthlib==3.2.0; python_version >= '3.6'
proto-plus==1.20.6; python_version >= '3.6'
protobuf==3.20.1; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydantic==1.9.1
pyparsing==3.0.9; python_version >= '3.1'
python-dotenv==0.20.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.28.1; python_version >= '3.7' and python_version < '4'
rsa==4.8; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.3.0; python_version >= '3.7'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn==0.18.2; python_version >= '3.7'
werkzeug==2.1.2; python_version >= '3.7'
yarl==1.7.2; python_version >= '3.6'
zipp==3.8.0; python_version >= '3.7'
